Move code for build-in optimizers out of __init__.py
I want to add more types of optimizers, that are based on optional
dependencies and are not built into sherpa. All those (built-in and optional)
optimizers shall be importable from `sherpa.optmethods`.
However, the optional optimizers will need to import some stuff from the
build-in ones. To avoid circular imports, it is thus better to move the build-in
optimizers into their own file instead of defining everything in `__init__.py`.
